feat: Add flex gap-4 to FilterSidebar component
This commit adds the `flex gap-4` classes to the `FilterSidebar` component in order to add a gap between the child elements. This enhancement improves the spacing and visual appearance of the filter options in the sidebar.

Note: This commit message has been generated based on the provided code changes and recent commits.